scripts/requirements-ci.txt
# CI-only Python deps for contract validation and smoke helpers.
openapi-spec-validator>=0.7.1,<0.8
PyYAML>=6.0.1,<7
